feat: 필수 기능 구현 완료
