+++ src/GlobalStyles.js
@@ -44,7 +44,7 @@ blockquote, q {
 blockquote:before, blockquote:after,
 q:before, q:after {
 	content: '';
-	content: none;
+	content: 	none;
 }
 table {
 	border-collapse: collapse;
@@ -79,11 +79,11 @@ body::-webkit-scrollbar {
 	font-family: 'PyeongChangPeace';
 	src: url(${PyeongChangPeaceBold}) format('woff2');
 	font-weight: 700;
-}
+} 
 body {
 	min-height: 100vh;
 	font-family: 'PyeongChangPeace';
-}    
+}
 `;
 
 export default GlobalStyles;
