+++ src/components/DetailLetter/index.jsx
@@ -1,13 +1,13 @@
 import React, { useState } from 'react';
-import { timestampToDate } from 'utils/date';
 import { useNavigate } from 'react-router-dom';
+import DetailLetterItem from 'components/DetailLetterItem';
 import * as S from './styles';
 
 export default function DetailLetter({ id, messages, setMessages }) {
-  const [isEdit, setIsEdit] = useState(false);
   const { nickname, content, createdAt, writedTo } = messages?.filter(
     (message) => message.id === id,
   )[0];
+  const [isEdit, setIsEdit] = useState(false);
   const [editContent, setEditContent] = useState(content);
   const navigate = useNavigate();
 
@@ -54,53 +54,18 @@ export default function DetailLetter({ id, messages, setMessages }) {
 
   return (
     <S.Detail>
-      <S.DetailContainer>
-        <S.DetailHeader>
-          <h2>NewJeans</h2>
-          <span>No. {id.slice(0, 8)}</span>
-        </S.DetailHeader>
-        <S.DetailInfo>
-          <S.DetailImg>
-            <img src={require(`assets/${writedTo}.jpg`)} alt="" />
-          </S.DetailImg>
-          <S.DetailDesc>
-            <div>
-              <h3>닉네임</h3>
-              <p>{nickname}</p>
-            </div>
-            <div>
-              <h3>작성일</h3>
-              <p>{timestampToDate(createdAt)}</p>
-            </div>
-            <div>
-              <h3>내용</h3>
-            </div>
-            <div>
-              {isEdit ? (
-                <textarea
-                  cols="30"
-                  rows="6"
-                  maxLength="100"
-                  value={editContent}
-                  onChange={handleChangeContent}
-                />
-              ) : (
-                <p>{content}</p>
-              )}
-            </div>
-          </S.DetailDesc>
-        </S.DetailInfo>
-        <S.DetailBtnBox>
-          <S.DetailBtn onClick={handleEdit} type={isEdit ? 'edit' : 'default'}>
-            {isEdit ? '완료' : '수정'}
-          </S.DetailBtn>
-          {!isEdit ? (
-            <S.DetailBtn onClick={handleDelete} type="delete">
-              삭제
-            </S.DetailBtn>
-          ) : undefined}
-        </S.DetailBtnBox>
-      </S.DetailContainer>
+      <DetailLetterItem
+        handleEdit={handleEdit}
+        handleDelete={handleDelete}
+        handleChangeContent={handleChangeContent}
+        nickname={nickname}
+        content={content}
+        createdAt={createdAt}
+        writedTo={writedTo}
+        isEdit={isEdit}
+        editContent={editContent}
+        id={id}
+      />
     </S.Detail>
   );
 }
